Develop

+++ src/main/scala/com/neighborhood/aka/laplace/estuary/web/utils/ValidationUtils.java
@@ -23,7 +23,7 @@ public class ValidationUtils {
     public static void notblank(String obj, String message) {
         Assert.isTrue(obj.trim() != "",message);
     }
-    public static void notZero(int num,String message) {
+    public static void notZero(long num,String message) {
         Assert.isTrue(num != 0,message);
     }
     public static void notEmpty(List obj, String message) {
